[qt-opengl] Add mesa/egl patches
See http://bugreports.qt.nokia.com/browse/QTBUG-9691 for discussion of
the issue and origin of these patches. This should fix compilation with
mesa >=7.8. It also adds a egl/no-egl configure switch, which I have
added a useflag for. I haven't looked into this to know which should be
preferred, but glx (no-egl) is the default.