[qt5-build.eclass] Run qmake "manually" at the top level too.
The recursive qmake that is executed at the end of configure is very
time-consuming, not parallelizable and especially it is completely useless,
since we already re-run qmake with custom arguments in all the relevant
directories. The only exception is the top-level project file (qtbase.pro),
which must be handled separately.

Therefore we can delete the recursive qmake invocation from configure
altogether. This also avoids pre-stripping of the qmake binary itself.