[qt4-build-multilib.eclass] Get rid of QT4_EXTRACT_DIRECTORIES.
When doing out-of-source builds, syncqt needs to be run. The problem
is that it insists on running on all Qt modules, even those that
have not been extracted, therefore it's bound to fail.

Since we're going to need out-of-source builds for multilib support
(we could use multilib_copy_sources but it'd nullify the benefits
of partial extraction), just remove the whole QT4_EXTRACT_DIRECTORIES
machinery now.

It was very difficult to maintain a precise list anyway...