[dev-qt/qtwebkit23] enable multilib support
This change enables proper multilib support so that 32 bit programs
depending on this, such as skype. Deps are forced to be either EAPI=5 or
multilib. Versions choosen are the one I tested with, mostly stable
packages for amd64 with exception of Qt libraries.

This also moves the builddir out of the source tree so there is no need to copy the
entire source folder for each ABI we are going to compile for.

The trick about using WEBKITOUTPUTDIR comes from fedora 21 SRPM, credit
to the maintainers for it.