[qt4-build-multilib] Disable instruction set when the corresponding flag is present
This wasn't necessary until now because the user's flags were overriding
the additional -mfoo flag passed from configure to compile.test. However
we will soon stop substituting the flags directly in compile.test because
that doesn't work in the multilib case, and there is no other feasible way
to have the user's flags take precedence over the appended ones during
the configure tests.

Therefore we now detect if the user disabled one or more instruction sets,
and pass the corresponding -no-foo option to configure.

Credit to kensington for the idea.