dev-qt/qtmultimedia: really fix egl/gles2 handling
The videonode plugin requires both egl and gles2, therefore depending
on [-gles2] is enough to disable the plugin, and makes much more sense
than depending on [-egl], since the USE flag is called 'gles2'.
Thanks to rhill for pointing this out.

While at it, use a [gles2=] dependency on both qtgui and qtwidgets,
like we do everywhere else in the Qt stack.

Gentoo-Bug: 576636
Package-Manager: portage-2.2.28