dev-qt: remove all 5.8.x ebuilds
The upstream 5.8 branch is basically dead (no more releases from it),
plus there are still several important breakages caused by the new
configuration system. Fixing 5.8.x is a waste of time at this point.
Let's focus our efforts on 5.9.0, which is expected to be released
at the end of the month.