HACK: Do not copy files that are missing in 5.9.9999
It seems that this one used to be a fix for the availability of qtdbus.
I've rebuild qtcore and qtdbus (and some other packages) with this
change active, and I can confirm that I can still build
mail-client/trojita which requires QtDBus...

And because there are no pre-5.9 ebuiolds anymore, the entire block can
be safely removed.