dev-qt/qtwebengine: remove USE="system-libvpx"
The patch does not apply at this time, without some changes, and
upstream has worked around the build issues that occur due to the
still unreleased version of 'libvpx' that is required.  As such, to
build using the bundled version, the flag is no longer needed, but
can be added back in the future if need be.

Package-Manager: Portage-2.3.51, Repoman-2.3.11