dev-qt/qtwebengine: fix has_version default-libcxx test for Clang
We moved the flag to sys-devel/clang-common to prepare for proper
config file support (which is backported to 15.0.2, although it's
going to be better in 16.x).

Go for the hack for now, but keep the bug open as we want to
switch to the proper link test from the LLVM ebuilds.

Bug: https://bugs.gentoo.org/875641
Bug: https://bugs.gentoo.org/875563
Signed-off-by: Sam James <sam@gentoo.org>